api/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
langgraph
google.generativeai
openai
python-dotenv